requirements.txt
torch
torchvision
gradio
Pillow
matplotlib
scikit-learn
tqdm
efficientnet-pytorch